Added Acoes FAB and RecyclerView and Adapter
1) Added the Floating Action Button on the Acoes Main Fragment. When
clicked, it will add a new testing stock value and a new CardView on the
RecyclerView
2) Added the RecyclerView on the Acoes Fragment to show the CardView of
the added Stocks. At first it will only show 1 stock, but clicked on the
FAB it will add one more at each click
3) Created the AcaoAdapter for the CardView to show the Stock at each
add.

Next Steps:
Change the layout of the CardView to show the Stock values as it should
and not only the company name and the ticker as the testing fase it is
as.