Created the Add New Stock DialogFragment
Created the add new stock DialogFragment when clicking on the FAB
button. Now when clicking on the FAB button, a DialogFragment will
appear for the user to input the new stock information. Then if
confirmed, the new stock will be shown in the portfolio, otherwise it
will dismiss the Dialog and nothing will happen.