Dialog not dismissed when validation fails
- Changed onPositiveButton to not dismiss Dialog always when click. It
will first validate inputted values on fields to check that it is not
empty. If all fields are correct, only then it will dismiss the Dialog,
send the intent and add a new stock
- Added raw validate functions of each field. For now they only check
that field is not empty. Later, functions will need to do Regex check to
validate if inputted values are consistent