Adding Fii Class, Adapter, Service, DialogFragment and layout
Adding Fii Class, Adapter, Service, DialogFragment and layout.
Now when selecting "Fundos Imobiliarios" on the navigation menu, it will
open to add a new fii to the portfolio. For now the code is very similar
to the Stock adding code, but when development progresses it will change
one from another.