Change from DialogFragment to normal form Fragment to add new investment
DialogFragment will not be used anymore to add a new investment when
clicking on FAB(stock, fii, etc). Now it will open a new activity with
the input form. The new activity will handle to add the inputted values
on database using provider and then finish() and return to MainActivity.