Back button is returning to wrong fragment in Add Stock Form bug
- Added some Logs to view flow of application
- There was a bug that when pressing the Up Navigation button of toolbar
on the AddFormActivity, it would return to wrong fragment of
MainActivity. This was happening because when calling back MainActivity,
it was recreating the instance of it and passing through onCreate again,
loading the first fragment presented. This was corrected by adding
"android:launchMode="singleTop" on the Manifest, that way the instance
is not recreated when returning to the MainActivity.