Added Stock Medium Price and isValidSellQuantity function
- Added Stock Medium Price on db to manage the medium price of stock
bought and sold by the user and use it to calculate total value on
portfolio.
- Create isValidSellQuantity that will check if it is a valid number
inserted and if there is enough of that stock in the portfolio to sell.
This will avoid user to sell a quanity of stock he dosent posses.
Example: User has 100 of ITUB4 and tries to sell 200 of ITUB4, this is
invalid.