Development of fragment to register Juros sobre Capital.
- Developed the register Juros sobre Capital by changing
DividendFormFragment to JCPDividendFormFragment. They both are almost
the same, so by passing the income_type by the intent we can do the
small difference between them.
- Created the function to calculate tax value and liquid receive value
to store on db when income_type is JCP. When income_type is Dividend,
tax value will be 0 and liquid and total receive value will be the same.
- Started development of IncomeDetailsActivity, to show details of the
income.