Improvement on Incomes sync
Improvement on Incomes sync.
Now it will pull the income at any position if it was not inserted yet.
This is done by comparing each of the retrieved income by the list of already inserted ones. If it has same per value, timestamp, type, then it is a already inserted income and should not insert again.